+++ src/app/cli/commands/todo/update/index.ts
@@ -1,0 +1,2 @@
+export * from './update-questions';
+export * from './update-todo.command';