+++ src/app/cli/commands/todo/update/update-questions.ts
@@ -1,0 +1,22 @@
+import {CliUtilityService, Question, QuestionSet} from "nest-commander";
+
+@QuestionSet({name: 'update-questions'})
+export class UpdateQuestions {
+  @Question({
+    message: 'Rename title todo:',
+    name: 'title',
+    type: 'input',
+  })
+  parseTitle(val: string) {
+    return val;
+  }
+
+  @Question({
+    message: 'Completed todo?',
+    name: 'completed',
+    type: 'confirm',
+  })
+  parseCompleted(val: string) {
+    return new CliUtilityService().parseBoolean(val.toLowerCase())
+  }
+}