+++ src/lib/todo/domain/entities/todo_completed.ts
@@ -1,6 +1,7 @@
-import { z } from 'zod';
+import {z} from 'zod';
 import {ValueObject} from "../../../shared";
-export class TodoCompleted extends ValueObject<boolean>{
+
+export class TodoCompleted extends ValueObject<boolean> {
   constructor(public readonly value: boolean) {
     super(value);
     this.ensureBool();