+++ test/todo/e2e/update-todo.command.spec.ts
@@ -1,6 +1,6 @@
 import {TestingModule} from '@nestjs/testing'
 import {AppModule, TodoDelete, TodoService, TodoUpdater} from "../../../src"
-import {TodoMother, TodoRepositoryMock} from "../stubs"
+import {TodoCompletedMother, TodoMother, TodoRepositoryMock, TodoTitleMother} from "../stubs"
 import {CommandTestFactory} from "nest-commander-testing"
 
 describe('Delete command', () => {
@@ -34,7 +34,12 @@ describe('Delete command', () => {
       .spyOn(todoUpdater, 'execute')
 
     // Act
-    await CommandTestFactory.run(commandInstance, ['update', todo1.todoId.value])
+    CommandTestFactory.setAnswers([
+      TodoTitleMother.random().value,
+      TodoCompletedMother.random().value ? 'y' : 'n'
+    ]);
+
+    await CommandTestFactory.run(commandInstance, ['todo', 'update', todo1.todoId.value])
 
     // Assert
     expect(processExit).toHaveBeenCalledWith(0)
@@ -50,15 +55,20 @@ describe('Delete command', () => {
       .spyOn(process, 'exit')
       .mockImplementation((code?: number) => undefined as never)
 
-    const todoCreatorMock = jest.spyOn(todoUpdater, 'execute')
+    const todoUpdaterMock = jest.spyOn(todoUpdater, 'execute')
 
     // Act
-    await CommandTestFactory.run(commandInstance, ['update', 'a'])
+    CommandTestFactory.setAnswers([
+      "a",
+      "o"
+    ]);
+
+    await CommandTestFactory.run(commandInstance, ['todo', 'update', 'a'])
 
     // Assert
     expect(processExit).toHaveBeenCalledWith(5)
     expect(todoRepositoryMock.saveMock).toHaveBeenCalledTimes(0)
-    expect(todoCreatorMock).toHaveBeenCalledTimes(0)
+    expect(todoUpdaterMock).toHaveBeenCalledTimes(0)
     expect(processExit).toHaveBeenCalledTimes(1)
     processExit.mockRestore()
   })